transcriber/requirements.txt
pika
python-dotenv
SQLAlchemy
pyodbc --no-binary=pyodbc
openai-whisper
minio
